Code refactoring
Split frontend functions into page specific groups
