+++ create.html
@@ -2,6 +2,7 @@
     <head>
         <script type="module">
             import {loadCreateVinyl} from "./js/ui-workflow.js";
+            
             loadCreateVinyl("new_vinyl", "vinyl_actions");
         </script>
     </head>
