+++ edit.html
@@ -2,6 +2,7 @@
     <head>
         <script type="module">
             import {loadEditVinyl} from "./js/ui-workflow.js";
+            
             loadEditVinyl("vinyl", "vinyl_actions");
         </script>
     </head>
