+++ index.html
@@ -1,8 +1,9 @@
 <html>
     <head>
         <script type="module">
-            import {loadVinylSearch} from "./js/ui-workflow.js";
-            loadVinylSearch("search_results", "search_filter");
+            import {loadSearchVinyl} from "./js/ui-workflow.js";
+            
+            loadSearchVinyl("search_results", "search_filter");
         </script>
     </head>
