+++ view.html
@@ -2,6 +2,7 @@
     <head>
         <script type="module">
             import {loadViewVinyl} from "./js/ui-workflow.js";
+            
             loadViewVinyl("vinyl", "vinyl_actions");
         </script>
     </head>
